commit changes of past 2 days, see changelog for details most important changes:  - added MSXPrinterPort  - PSG sounds louder
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@620 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/AY8910.hh
@@ -39,7 +39,7 @@ class AY8910 : public SoundDevice
 		void reset();
 		void setVolume(short newVolume);
 		void setSampleRate(int sampleRate);
-		short* updateBuffer(int length);
+		int* updateBuffer(int length);
 		
 	private:
 		void wrtReg(byte reg, byte value);
@@ -80,7 +80,7 @@ class AY8910 : public SoundDevice
 		unsigned int volTable[16];
 		byte oldEnable;
 		
-		short* buffer;
+		int* buffer;
 		
 		AY8910Interface &interface;
 };
